requirements.txt
attrs==23.1.0
click==8.1.3
terminaltables==3.1.10
PyYAML==6.0.0
tzdata==2023.3
pymongo==4.3.3
